requirements.txt
# 基础依赖
pyaudio==0.2.11
pygame==2.5.2
numpy==1.26.4
librosa==0.10.1
soundfile==0.12.1

# Whisper语音识别
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2

# 语音合成
aiohttp==3.9.1
websockets==12.0

# 数据库支持
mysql-connector-python==8.1.0
pandas==2.0.3

# 自然语言处理
jieba==0.42.1
scikit-learn==1.3.0

# 其他工具
requests==2.31.0

# 语音合成服务调用
requests>=2.25.0

# 音频处理
pygame>=2.0.0
pyaudio>=0.2.11

# 语音识别
vosk>=0.3.45

# Web框架
flask>=2.0.0
flask-socketio>=5.0.0
flask-cors>=3.0.0
